finalize runtests splitting
