+++ test/runtests.jl
@@ -1,25 +1,19 @@
-using BeamletOptics
-using Test
-using LinearAlgebra
-using GeometryBasics
-using AbstractTrees
-
-const BMO = BeamletOptics
-
-const mm = 1e-3
-const inch = 25.4mm
-
 # order or inclusion matters!
-
 include(joinpath(@__DIR__, "TestUtils.jl"))
 include(joinpath(@__DIR__, "TestAbstractTypes.jl"))
+
+# Test ray and beam types
 include(joinpath(@__DIR__, "TestRays.jl"))
+include(joinpath(@__DIR__, "TestPolarizedRays.jl"))
 include(joinpath(@__DIR__, "TestBeams.jl"))
 include(joinpath(@__DIR__, "TestBeamGroups.jl"))
+include(joinpath(@__DIR__, "TestGaussianBeamlet.jl"))
 
 # Test geometry representation
 include(joinpath(@__DIR__, "Geometry", "TestMesh.jl"))
 include(joinpath(@__DIR__, "Geometry", "TestSDFs.jl"))
+
+# Test system and object containers
 include(joinpath(@__DIR__, "TestSystem.jl"))
 include(joinpath(@__DIR__, "TestObjectGroups.jl"))
 
@@ -29,358 +23,18 @@ include(joinpath(@__DIR__, "Lenses", "TestSurfaces.jl"))
 include(joinpath(@__DIR__, "Lenses", "TestAsphericalLenses.jl"))
 include(joinpath(@__DIR__, "Lenses", "TestCylindricalLenses.jl"))
 
-include(joinpath(@__DIR__, "TestGaussianBeamlet.jl"))
-
 # Test component models
 include(joinpath(@__DIR__, "Components", "TestDummies.jl"))
 include(joinpath(@__DIR__, "Components", "TestDetector.jl"))
 include(joinpath(@__DIR__, "Components", "TestBeamsplitters.jl"))
+include(joinpath(@__DIR__, "Components", "TestPolarizers.jl"))
 
 # Test end-to-end models
 include(joinpath(@__DIR__, "E2E", "TestDoubleGaussLens.jl"))
 include(joinpath(@__DIR__, "E2E", "TestMichelson.jl"))
+include(joinpath(@__DIR__, "E2E", "TestMachZehnder.jl"))
 
 # Test regressions
 include(joinpath(@__DIR__, "TestBugFixes.jl"))
 
-@testset "Polarized rays" begin
-    @testset "Polarization transforms" begin
-        # Reflection matrix and lin. x-pol
-        J = BMO.SPBasis(-1, 0, 0, 1)
-        E0 = [1, 0, 0]
-
-        @testset "90° reflection" begin
-            in_dir = [0, 0, 1]
-            out_dir = [1, 0, 0]
-            nml = normalize([1, 0, -1])
-            P90 = BMO._calculate_global_E0(in_dir, out_dir, nml, J)
-            @test P90 * E0 ≈ [0, 0, -1]
-            @test P90 * in_dir ≈ out_dir
-        end
-
-        @testset "0° reflection" begin
-            in_dir = [0, 0, 1]
-            out_dir = [0, 0, -1]
-            nml = normalize([0, 0, -1])
-            P00 = BMO._calculate_global_E0(in_dir, out_dir, nml, J)
-            @test P00 * E0 ≈ [-1, 0, 0]
-            @test P00 * in_dir ≈ out_dir
-        end
-    end
-
-    @testset "Test error messages" begin
-        # Test dir. error msg
-        @test_throws ErrorException PolarizedRay(zeros(3), zeros(3))
-        @test_throws ErrorException PolarizedRay(zeros(3), ones(3)*eps())
-        # Test polarization orthogonal error msg
-        @test_throws ErrorException PolarizedRay(zeros(3), [0,1,0], 1e-6, [0,1,1])
-    end
-
-    @testset "Mirror reflections" begin
-        # Setup system as in https://opg.optica.org/ao/fulltext.cfm?uri=ao-50-18-2855&id=218813
-        m1 = SquarePlanoMirror2D(1.0)
-        m2 = SquarePlanoMirror2D(1.0)
-        m3 = SquarePlanoMirror2D(1.0)
-        translate3d!(m2, [2, 0, 0])
-        translate3d!(m3, [2, 2, 0])
-        zrotate3d!(m1, deg2rad(-90))
-        yrotate3d!(m1, deg2rad(45))
-        zrotate3d!(m2, deg2rad(45))
-        xrotate3d!(m3, deg2rad(135))
-
-        system = StaticSystem([m1, m2, m3])
-
-        I0_1 = 1
-        I0_2 = 5
-        lin_x_pol = [I0_1, 0, 0]
-        lin_y_pol = [0, I0_2, 0]
-
-        # Beam of polarized rays
-        ray = PolarizedRay([0.0, 0, -2], [0, 0, 1], 1000e-9, lin_x_pol)
-        beam = Beam(ray)
-
-        @testset "x-Polarization" begin
-            BMO.polarization!(ray, lin_x_pol)
-            # test tracing
-            solve_system!(system, beam)
-            @test BMO.polarization(beam.rays[1]) ≈ lin_x_pol
-            @test BMO.polarization(beam.rays[2]) ≈ [0, 0, -I0_1]
-            @test BMO.polarization(beam.rays[3]) ≈ [0, 0, I0_1]
-            @test BMO.polarization(beam.rays[4]) ≈ [0, -I0_1, 0]
-            @test length(beam) == 6.0
-        end
-
-        @testset "y-Polarization" begin
-            BMO.polarization!(ray, lin_y_pol)
-            translate3d!(m3, [0, 2, 0])
-            # test retracing
-            solve_system!(system, beam)
-            @test BMO.polarization(beam.rays[1]) ≈ lin_y_pol
-            @test BMO.polarization(beam.rays[2]) ≈ [0, -I0_2, 0]
-            @test BMO.polarization(beam.rays[3]) ≈ [I0_2, 0, 0]
-            @test BMO.polarization(beam.rays[4]) ≈ [-I0_2, 0, 0]
-            @test length(beam) == 8.0
-        end
-    end
-
-    @testset "Brewster windows" begin
-        brewster_angle(n) = atan(n)
-        # Testcase based on 5 successive Brewster windows
-        n = 1.5
-        θb = brewster_angle(n)
-        d = 0.1
-        # Calculate transmission efficiency
-        rs, rp, ts, tp = BMO.fresnel_coefficients(θb, n)
-        Ts = 1 - abs2(rs)
-        Tp = 1 - abs2(rp)
-        # Setup testcase
-        s1 = BMO.CuboidMesh(1.0, d, 1.0)
-        s2 = BMO.CuboidMesh(1.0, d, 1.0)
-        s3 = BMO.CuboidMesh(1.0, d, 1.0)
-        s4 = BMO.CuboidMesh(1.0, d, 1.0)
-        s5 = BMO.CuboidMesh(1.0, d, 1.0)
-        l1 = Lens(s1, x -> n)
-        l2 = Lens(s2, x -> n)
-        l3 = Lens(s3, x -> n)
-        l4 = Lens(s4, x -> n)
-        l5 = Lens(s5, x -> n)
-        translate3d!.([l1, l2, l3, l4, l5], Ref([-0.5, -d / 2, -0.5]))
-        BMO.set_new_origin3d!.(BMO.shape.([l1, l2, l3, l4, l5]))
-        translate3d!(l2, [0, 0.5, -1d / 2])
-        translate3d!(l3, [0, 1.0, -2d / 2])
-        translate3d!(l4, [0, 1.5, -3d / 2])
-        translate3d!(l5, [0, 2.0, -4d / 2])
-        xrotate3d!.([l1, l2, l3, l4, l5], -θb)
-        # Solve system of s- and p-polarized beams
-        system = StaticSystem([l1, l2, l3, l4, l5])
-        x_pol_ray = PolarizedRay(
-            [-0.1, -1, 0], [0, 1.0, 0], 1000e-9, [BMO.electric_field(1), 0, 0])
-        z_pol_ray = PolarizedRay(
-            [+0.1, -1, 0], [0, 1.0, 0], 1000e-9, [0, 0, BMO.electric_field(1)])
-        s_beam = Beam(x_pol_ray)
-        p_beam = Beam(z_pol_ray)
-        solve_system!(system, s_beam)
-        solve_system!(system, p_beam)
-        # Since system is non-focussing, calculate pseudo-intensity
-        pseudo_Is = abs2(BMO.polarization(last(BMO.rays(s_beam)))[1]) /
-                    (2 * BMO.Z_vacuum)
-        pseudo_Ip = abs2(BMO.polarization(last(BMO.rays(p_beam)))[3]) /
-                    (2 * BMO.Z_vacuum)
-        # Test against m interfaces
-        m = length(system.objects) * 2
-        @test pseudo_Is ≈ Ts^m
-        @test pseudo_Ip ≈ Tp^m
-    end
-
-    @testset "Fresnel rhomb" begin
-        # Create Fresnel rhomb with n=1.5 and θ=53.3° for quarter-wave plate effect
-        n = 1.5
-        s1 = BMO.CuboidMesh(0.5, 1.25, 0.5, deg2rad(53.3))
-        l1 = Lens(s1, x -> n)
-        translate3d!(l1, [-0.25, 0, -0.25])
-        BMO.set_new_origin3d!(s1)
-        # Rotate prism to obtain 45° beam input polarization
-        yrotate3d!(l1, deg2rad(135))
-        # Solve system
-        system = StaticSystem([l1])
-        ray = PolarizedRay(
-            [0, -1, 0], [0, 1.0, 0], 1000e-9, [0, 0, BMO.electric_field(1)])
-        beam = Beam(ray)
-        solve_system!(system, beam)
-        # Assumes propagation along the y-axis after rhomb, calculate polarization state
-        Ex = getindex.(BMO.polarization.(beam.rays), 1)
-        Ey = getindex.(BMO.polarization.(beam.rays), 2)
-        Ez = getindex.(BMO.polarization.(beam.rays), 3)
-        # Test for circular polarization and Ey error
-        phi = angle(last(Ez)) - angle(last(Ex))
-        @test phi ≈ π / 2
-        @test abs(last(Ey)) < 2e-14
-    end
-
-    @testset "Mach-Zehnder Interferometer" begin
-        # setup MZI
-        m1 = SquarePlanoMirror2D(BMO.inch)
-        m2 = SquarePlanoMirror2D(BMO.inch)
-        b1 = ThinBeamsplitter(BMO.inch, reflectance = 0.5)
-        b2 = ThinBeamsplitter(BMO.inch, reflectance = 0.5)
-
-        system = StaticSystem([m1, m2, b1, b2])
-
-        translate3d!(b1, [0 * BMO.inch, 0 * BMO.inch, 0])
-        translate3d!(b2, [2 * BMO.inch, 2 * BMO.inch, 0])
-        translate3d!(m1, [0 * BMO.inch, 2 * BMO.inch, 0])
-        translate3d!(m2, [2 * BMO.inch, 0 * BMO.inch, 0])
-
-        # Rotate with consideration to mirror/bs normal
-        zrotate3d!(b1, deg2rad(360 - 135))
-        zrotate3d!(b2, deg2rad(45))
-        zrotate3d!(m1, deg2rad(360 - 135))
-        zrotate3d!(m2, deg2rad(45))
-
-        ray = PolarizedRay([0, -0.1, 0], [0.0, 1.0, 0], 1000e-9, [0, 0, 1])
-        beam = Beam(ray)
-
-        @testset "z-polarized ray along y-axis" begin
-            # Solve with z-polarized ray along y-axis
-            BMO.polarization!(ray, [0, 0, 1])
-            solve_system!(system, beam)
-
-            # Extract E0s: t - transmitted, r - reflected
-            t = BMO.polarization(beam.children[1].rays[1])
-            r = BMO.polarization(beam.children[2].rays[1])
-            tr = BMO.polarization(beam.children[1].rays[2])
-            rr = BMO.polarization(beam.children[2].rays[2])
-            trt = BMO.polarization(beam.children[1].children[1].rays[1])
-            trr = BMO.polarization(beam.children[1].children[2].rays[1])
-            rrt = BMO.polarization(beam.children[2].children[1].rays[1])
-            rrr = BMO.polarization(beam.children[2].children[2].rays[1])
-
-            # Test phase flips
-            @test t[3] ≈ sqrt(2) / 2
-            @test r[3] ≈ -sqrt(2) / 2
-            @test tr[3] ≈ -sqrt(2) / 2
-            @test rr[3] ≈ sqrt(2) / 2
-            @test trt ≈ rrr
-            @test trr ≈ rrt
-        end
-
-        @testset "x-polarized ray along y-axis" begin
-            # Test num. of leaves before retracing
-            @test length(collect(Leaves(beam))) == 4
-            # Retrace with z-polarized ray along y-axis
-            BMO.polarization!(ray, [1, 0, 0])
-            solve_system!(system, beam)
-
-            # Extract E0s: t - transmitted, r - reflected
-            t = BMO.polarization(beam.children[1].rays[1])
-            r = BMO.polarization(beam.children[2].rays[1])
-            tr = BMO.polarization(beam.children[1].rays[2])
-            rr = BMO.polarization(beam.children[2].rays[2])
-            trt = BMO.polarization(beam.children[1].children[1].rays[1])
-            trr = BMO.polarization(beam.children[1].children[2].rays[1])
-            rrt = BMO.polarization(beam.children[2].children[1].rays[1])
-            rrr = BMO.polarization(beam.children[2].children[2].rays[1])
-
-            # Test phase flips
-            @test t[1] ≈ sqrt(2) / 2
-            @test r[2] ≈ -sqrt(2) / 2
-            @test tr ≈ r
-            @test rr ≈ t
-            @test trt ≈ rrr
-            @test trr ≈ rrt
-        end
-    end
-end
-
-@testset "Polarizing components" begin
-    # Test fcts.
-    pseudo_I(v) = norm(v)^2
-    pseudo_I(ray::BMO.PolarizedRay) = pseudo_I(BMO.polarization(ray))
-    malus_law(θ) = cosd(θ)^2
-    function generate_linpol_angles(thetas)
-        angles = zeros(length(thetas))
-        # consider flip in angle quadrant above 90°  
-        for i in eachindex(thetas)
-            theta = thetas[i] - 1
-            if theta > 90 && theta <= 270
-                angles[i] = 180
-            end
-        end
-        return angles
-    end
-
-    @testset "Jones polarization matrices" begin
-        @test isdefined(BMO, :AbstractJonesMatrix)
-        @test isdefined(BMO, :LocalJonesBasis)
-        @test isdefined(BMO, :GlobalJonesBasis)
-        # test matrix operations
-        J = BMO.XYBasis(1, 2, 3, 4)
-        R = [1 2 0; 3 4 0; 0 0 1]
-        @test J == R
-        @test J*2 == R*2
-        @test transpose(J) == transpose(R)
-        @test inv(J) ≈ inv(R)
-    end
-
-    @testset "Polarizing filter" begin
-        filter = PolarizationFilter(5mm)
-        system = System([filter])
-        # rotation angle steps, intensity and angle comparison
-        thetas = 1:10:360
-        i_n = zeros(length(thetas))
-        i_a = similar(i_n)
-        angles_n = similar(i_n)
-        angles_a = generate_linpol_angles(thetas)
-
-        @testset "Pol. filter normal incidence - rotation around y-axis" begin 
-            Rfil = orientation(filter)
-            ray_dir = Rfil[:,2]
-            ray_pos = position(filter) - 10mm * ray_dir
-            local_x = Rfil[:,1]
-            local_z = Rfil[:,3]
-            pol_vec = local_x
-            lambda = 1e-6
-            beam = Beam(ray_pos, ray_dir, lambda, pol_vec)            
-            # step ref. vector by angle increment 
-            RotMat = BMO.rotate3d(ray_dir, deg2rad(step(thetas)))
-            for i in eachindex(thetas)
-                solve_system!(system, beam)
-                E1 = BMO.polarization(BMO.rays(beam)[2])
-                i_n[i] = pseudo_I(E1)
-                i_a[i] = malus_law(thetas[i]-1)
-                rotate3d!(filter, ray_dir, deg2rad(step(thetas)))
-                # test polarization state
-                @test all(BMO.islinear.(beam.rays))
-                # test projected polarization direction
-                v1 = real(E1)
-                angles_n[i] = rad2deg(BMO.angle3d(v1, pol_vec))
-                # update ref vector
-                pol_vec = RotMat * pol_vec
-            end    
-            @test angles_n ≈ angles_a
-        end
-
-        # Move and rotate filter
-        translate3d!(filter, [0,10mm,0])
-        xrotate3d!(filter, deg2rad(45))
-        zrotate3d!(filter, deg2rad(30))
-
-        @testset "Pol. filter tilted incidence - rotation around ray optical axis" begin
-            Rfil = orientation(filter)
-            ray_dir = Rfil[:,2]
-            ray_pos = position(filter) - 10mm * ray_dir
-            local_x = Rfil[:,1]
-            local_z = Rfil[:,3]
-            pol_vec = local_x
-            lambda = 1e-6
-            beam = Beam(ray_pos, ray_dir, lambda, pol_vec)
-            # tilt filter
-            θ_tilt = 45
-            rotate3d!(filter, local_x, deg2rad(θ_tilt))
-            i_n_tilt = similar(i_n)
-            for i in eachindex(thetas)
-                solve_system!(system, beam)
-                i_n_tilt[i] = pseudo_I(beam.rays[2].E0)
-                rotate3d!(filter, ray_dir, deg2rad(step(thetas)))
-                @test all(BMO.islinear.(beam.rays))
-            end
-            # only applicable for θ_tilt = 45°
-            i_a_tilt = malus_law.(thetas .- 1) .* 0.75 .+ 0.25
-            @test i_n_tilt ≈ i_a_tilt
-        end
-    end
-end
-
-@testset "Render" begin
-    axis = nothing
-    cube = BMO.CubeMesh(1)
-    @test_throws BMO.MissingBackendError render!(axis, cube)
-    @test_throws BMO.MissingBackendError BMO.get_view(axis)
-    @test_throws BMO.MissingBackendError BMO.set_view(axis, [1 1; 0 0])
-    @test_throws BMO.MissingBackendError BMO.hide_axis(axis, true)
-end
-
-@testset "Aqua" begin
-    using Aqua
-    Aqua.test_all(BMO)
-end
+include(joinpath(@__DIR__, "TestMisc.jl"))
